# change: [M3-9997] - Update estimated time for LKE-E node pool pending creation message