--- a/packages/manager/.changeset/pr-12251-upcoming-features-1747763494658.md
+++ b/packages/manager/.changeset/pr-12251-upcoming-features-1747763494658.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Upcoming Features
+---
+
+Update estimated time for LKE-E node pool pending creation message ([#12251](https://github.com/linode/manager/pull/12251))