## Description 📝

As reported by LKE QA, the amount of time it takes to fully provision LKE-E cluster nodes has increased since we originally determined that nodes were provisioned within 10 minutes or less. (On staging, it takes ~1110s, or ~18 minutes.)

This PR updates the util and copy to display this message for the first 20 minutes after cluster creation.

## Changes  🔄
- Renamed the util for `20` min and updated the interval
- Updated the copy to mention 20 minutes
- Updated test coverage

## Target release date 🗓️

6/3

## Preview 📷

| Before  | After   |
| ------- | ------- |
| ![Screenshot 2025-05-20 at 10 41 39 AM](https://github.com/user-attachments/assets/9e092420-e7ad-4a9a-961b-ac0020f13919) | ![Screenshot 2025-05-20 at 10 41 22 AM](https://github.com/user-attachments/assets/de4b66ef-ff61-447c-9dd4-4c3b7b1f63eb)|

## How to test 🧪

### Prerequisites

(How to setup test environment)

- Have the LKE-E customer tag on your account or use `prod-test-004-fe` using our shared credentials
- Have the LKE-E feature flag on

### Reproduction steps

(How to reproduce the issue, if applicable)

- [ ] Check out `develop` and create a new LKE-E cluster
- [ ] Observe the node pools section has a placeholder message stating 10 minutes
- [ ] Observe that if the cluster is not ready after 10 minutes, the message disappears and the empty table state message is displayed

### Verification steps

(How to verify changes)

- [ ] Check out this branch and create a new LKE-E cluster
- [ ] Observe that after 10 minutes and up to 20 minutes, the placeholder message displays, stating 20 minutes
- [ ] Observe that the message continues to display until your node pools appear - the idea is that 20 minutes should generally be enough time - let me know if that's not the case

<details>
<summary> Author Checklists </summary>

## As an Author, to speed up the review process, I considered 🤔

👀 Doing a self review
❔ Our [contribution guidelines](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md)
🤏 Splitting feature into small PRs
➕ Adding a [changeset](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md#writing-a-changeset)
🧪 Providing/improving test coverage
 🔐 Removing all sensitive information from the code and PR description
🚩 Using a feature flag to protect the release
👣 Providing comprehensive reproduction steps
📑 Providing or updating our documentation
🕛 Scheduling a pair reviewing session
📱 Providing mobile support
♿  Providing accessibility support

<br/>

- [ ] I have read and considered all applicable items listed above.

## As an Author, before moving this PR from Draft to Open, I confirmed ✅

- [ ] All unit tests are passing
- [ ] TypeScript compilation succeeded without errors
- [ ] Code passes all linting rules

</details>